Use the schema2ldif tool to convert .schema files to .ldif
This avoids the need for the fusiondirectory-insert-schema tool, which should
be the last piece of the puzzle to support slapd in Bookworm.

Note that schema2ldif has been available since Buster at least and that
it was a dependency of the previous schema packages, so it's not actually
a new requirement.