Update default tasks on intial installation
First, the role tries to create the cn=Hidden Object Viewer,ou=Roles,...
entry with a "memberOf" attribute pointing to the
cn=Hidden Objects, ou=Groups,... group (which is created in the next task).

This causes the OpenLDAP version in Bookworm to complain (edited to break
up overlong lines):

TASK [debops.debops.slapd : Create cn=Hidden Object Viewer role] ***************
An exception occurred during task execution. To see the full traceback, use -vvv.
The error was: ldap.CONSTRAINT_VIOLATION: {
	'msgtype': 105,
	'msgid': 4,
	'result': 19,
	'desc': 'Constraint violation',
	'ctrls': [],
	'info': 'memberOf: no user modification allowed'}

Ok, so removing the "memberOf" attribute allows the role to continue to the
next default task, which is the actual creation of the
cn=Hidden Objects, ou=Groups,... group...which has a similar issue, it tries
to set both the "member" and "memberOf" attribute at the same time...which
explodes in a similar manner.

TASK [debops.debops.slapd : Create cn=Hidden Objects group] ********************
An exception occurred during task execution. To see the full traceback, use -vvv.
The error was: ldap.CONSTRAINT_VIOLATION: {
	'msgtype': 105,
	[snip]
	'info': 'memberOf: no user modification allowed'}

First, I tried just removing the "memberOf" attributes while keeping the
"member" attribute intact, but that meant that cn=Hidden Objects did not
get a "memberOf" attribute. It seems that OpenLDAP 2.5 is stricter in this
sense.

So the approach I took was for the the cn=Hidden Objects entry
to be created first, and then added to itself in a subsequent operation
in order for the entry to look like it did in Bullseye and earlier.

I'm hoping this is backwards compatible, but it works on Bookworm.